+++ app/controllers/index.js
@@ -1,7 +1,7 @@
 'use strict'
 const controllers = {}
 controllers.users = require('./users')
-controllers.articleCategory = require('./article-category')
+controllers.category = require('./category')
 controllers.article = require('./article')
 
 module.exports = controllers