+++ app/models/index.js
@@ -5,6 +5,6 @@ mdbs.User = require('./users')
 mdbs.Role = require('./role')
 mdbs.Right = require('./right')
 mdbs.Article = require('./article')
-mdbs.ArticleCategory = require('./article-category')
+mdbs.ArticleCategory = require('./category')
 
 module.exports = mdbs