+++ app/services/index.js
@@ -2,7 +2,7 @@
 const Services = {}
 
 Services.users = require('./users')
-Services.articleCategory = require('./article-category')
+Services.category = require('./category')
 Services.article = require('./article')
 
 module.exports = Services