+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-* **article-category:** 文章分类，建表和创建的接口 ([35d0b59](https://github.com/morehao/express-restfulApi/commit/35d0b59))
+* **category:** 文章分类，建表和创建的接口 ([35d0b59](https://github.com/morehao/express-restfulApi/commit/35d0b59))
 * **base-service:** baseservice增加findById，家、公司都可搬砖 ([89808cb](https://github.com/morehao/express-restfulApi/commit/89808cb))
 * **logs:** 优化日志，格式化日志的保存与输出 ([a909324](https://github.com/morehao/express-restfulApi/commit/a909324))